Change controller_block_overhead to quantity
When specifying resources like disk sizes, the value is expected to
represent a quantity rather than an integer so users can set values like
"1Gi". Applying this to controller_block_overhead as well.

+++ pkg/controller/plan/util/utils.go
@@ -27,7 +27,7 @@ func CalculateSpaceWithOverhead(requestedSpace int64, volumeMode *core.Persisten
 	if *volumeMode == core.PersistentVolumeFilesystem {
 		spaceWithOverhead = int64(math.Ceil(float64(alignedSize) / (1 - float64(settings.Settings.FileSystemOverhead)/100)))
 	} else {
-		spaceWithOverhead = alignedSize + int64(settings.Settings.BlockOverhead)
+		spaceWithOverhead = alignedSize + settings.Settings.BlockOverhead
 	}
 	return spaceWithOverhead
 }
